requirements.txt
pytest
allure-pytest
requests
pyyaml
pycryptodome
pytest-ordering
pymysql
Flask
gunicorn
